+++ handlers/category.py
@@ -1,6 +1,6 @@
 import flask
 from google.appengine.ext import ndb
-from werkzeug.exceptions import BadRequest
+from werkzeug.exceptions import BadRequest, Unauthorized
 
 import models
 import login
@@ -9,16 +9,17 @@ from itemcatalog import app
 @app.route('/catalog/<catalog_id>/addcategory', methods=['GET', 'POST'])
 @login.check_logged_in
 def add_category_handler(catalog_id):
-    username = flask.session.get('username')
+    user = models.get_current_user()
     catalog_id = int(catalog_id)
     catalog_entity = models.get_catalog_by_id(catalog_id)
     if not catalog_entity:
         raise BadRequest('Could not find catalog with id %d' % catalog_id)
 
-    models.check_user_owns(catalog_entity)
+    if not catalog_entity.user_can_edit(user):
+        raise Unauthorized
 
     if flask.request.method == 'GET':
-        return flask.render_template('addcategory.html', username=username, catalog=catalog_entity, category=None)
+        return flask.render_template('addcategory.html', user=user, catalog=catalog_entity, category=None)
 
     elif flask.request.method == 'POST':
         new_category = models.Category(name=flask.request.form.get('name'),
@@ -31,13 +32,14 @@ def add_category_handler(catalog_id):
 @app.route('/catalog/<catalog_id>/editcategory/<category_id>', methods=['GET', 'POST'])
 @login.check_logged_in
 def edit_category_handler(catalog_id, category_id):
-    username = flask.session.get('username')
+    user = models.get_current_user()
     catalog_id = int(catalog_id)
     catalog_entity = models.get_catalog_by_id(catalog_id)
     if not catalog_entity:
         raise BadRequest('Could not find catalog with id %d' % catalog_id)
 
-    models.check_user_owns(catalog_entity)
+    if not catalog_entity.user_can_edit(user):
+        raise Unauthorized
 
     category_id = int(category_id)
     category_entity = models.get_category_by_id(catalog_id, category_id)
@@ -45,7 +47,7 @@ def edit_category_handler(catalog_id, category_id):
         raise BadRequest('Could not find category with id %d!' % category_id)
 
     if flask.request.method == 'GET':
-        return flask.render_template('editcategory.html', username=username, catalog=catalog_entity, category=category_entity)
+        return flask.render_template('editcategory.html', user=user, catalog=catalog_entity, category=category_entity)
 
     elif flask.request.method == 'POST':
         category_entity.name = flask.request.form.get('name')
@@ -57,7 +59,6 @@ def edit_category_handler(catalog_id, category_id):
 @app.route('/catalog/<catalog_id>/deletecategory/<category_id>', methods=['POST'])
 @login.check_logged_in
 def delete_category_handler(catalog_id, category_id):
-    username = flask.session.get('username')
     catalog_id = int(catalog_id)
     catalog_entity = models.get_catalog_by_id(catalog_id)
     if not catalog_entity:
@@ -67,6 +68,9 @@ def delete_category_handler(catalog_id, category_id):
     category_entity = models.get_category_by_id(catalog_id, category_id)
     if not (category_entity):
         raise BadRequest('Could not find category with id %d!' % category_id)
+
+    if not catalog_entity.user_can_edit(models.get_current_user()):
+        raise Unauthorized
 
     models.delete_category(catalog_id, category_id)
     models.wait_for(category_entity)
